fix(jdbc): ignore invalid sqlserver URLs
